Move the storage checker from pyanaconda.storage
Move the storage checker and the validation checks to the Storage module.
