+++ pyanaconda/modules/storage/checker/utils.py
