+++ pyanaconda/modules/storage/partitioning/interactive/interactive_partitioning.py
@@ -23,7 +23,7 @@ from pyanaconda.modules.storage.bootloader.execution import setup_bootloader
 from pyanaconda.modules.storage.partitioning.automatic.automatic_partitioning import \
     AutomaticPartitioningTask
 from pyanaconda.modules.storage.partitioning.base_partitioning import PartitioningTask
-from pyanaconda.storage.checker import storage_checker, verify_luks_devices_have_key
+from pyanaconda.modules.storage.checker.utils import storage_checker, verify_luks_devices_have_key
 
 log = get_module_logger(__name__)
