+++ tests/nosetests/pyanaconda_tests/simple_import_test.py
@@ -71,8 +71,8 @@ class SimpleImportTestCase(unittest.TestCase):
             "pyanaconda.core.configuration.anaconda",
             "pyanaconda.isys",
             "pyanaconda.modules.common.constants.interfaces",
+            "pyanaconda.modules.storage.checker.utils",
             "pyanaconda.payload.source.factory",
-            "pyanaconda.storage.checker",
             "pyanaconda.ui.categories",
             "pyanaconda.ui.gui.spokes.lib.cart",
             "pyanaconda.ui.tui.spokes.askvnc",
